feat: change SwapPreviw to SwapThumbnailPreview
